requirements.txt
fastapi
uvicorn
faiss-cpu
transformers
openai
requests
